Add perl6-b binary target
I've wanted a binary runner for perl6 for a long time,
starting with the need to sometimes run the thing under
gdb. If we want to be fully 'production ready', I think
this should be a necessity too.

Note that this does not currently make a 'fake-excutable'
(like pbc2exe) does, rather relying on the installation
of moarvm files in their respective directories. Since we're
likely to want a 'static' perl6 binary eventually, this should
be investigated.

Also note that it uses hardcoded moarvm include dependencies,
which isn't really necessary since nqp --show-config can
provide us with this information.